RowBinary serializer and related additions and fixes
Still TODO: a way to avoid regenerating the same function for every
fields. Currently, the `oc` parameters can not be factored out although
it's essentially always the same, forcing the functions to print again
the cond and loop functions they need in a potentially new `oc` (but in
actuality constant).
So, the output API has to be changed somehow.