Consider user types equal to their definition in all occurrences
Makes testing des/ser much simpler, since this information is lost
during des/ser.

We could have two version of the comparison operator though, one that
would apply when comparing a user type and another type is a single
expression, so there is a switch to control this behavior, but default
is to consider them equal for now.