Add an apply operation
So far functions could only be used to parameterize loop operations. Now
user programs can make use of them too.

The motivation was to provide a clean interface for supplying the transform
parameter of the dessser function with a use provided function (that is
obtained from an s-expression on the command line).