First implementation of a TOP
TOP is here implemented as a special type of set. The new "insert_weighted"
operator is required to specify the weight in addition to the item value,
as requiring that the weight be part of the value would require an
extractor for the weight, for the value, and also for the time, and to
perform weight inflation (ie. decay) in the top itself.
On the other hand, with an explicit weight it is now easy to perform decay
from the outside.

insert_weighted still TODO.