Proof of concept for recursivity
`T.This` is a new type, designating the toplevel defined type.
(De)serializing a type containing such self references requires that
the des/ser functions be recursive.
Therefore, `E.Myself` is a new expression that refers to the innermost
function, so it can be called without knowing its name.

Only implemented in HeapValue.serialize for now. This also require this
function to change type: instead of returning an expression that serialize
it must now return an expression that defines a function that serialize
(so that this function can call itself recursively).

Oh, and none of the backend actually do implement any of that.