Do away with the `Data` type
Ref had to be at the bottom of the type hierarchy so that any type could
be mutable. But it's also desirable to have vectors/records of Refs,
despite Refs are not serializable.

So now the type DessserTypes.t does not distinguish between serializable
and non serializable types (for now). It just raises when the user ask
to serialize something that's not serializable (like, say, a pointer or
a function).

Also, this will allow to simplify the types by removing the special "low
level" types (bit, byte, words... and pairs).