Do away with special type for pairs (replaced with tuples)
The hard work was to make Dump work for C++: previously, we had an
operator<< for Pair, Pointer and Sizes which was good enough but now that
pairs are tuples we need an operator<< for tuples, created for any defined
tuples, ie. requiring that an operator<< be defined for any possible
type.