Do not pass environment everywhere (only to type check, eval and generate)
This makes the code lighter and also helps the JSON parser that can now
manipulate the skip function before it's properly created in the compilation
unit.

The price to pay is that it is no longer possible to get any expression's
type. Instead, some StdLib functions now requires the type of some
parameters to be explicitly given. Not a big hassle in practice.
The other consequence is that NotNull must now accept already nullable
arguments. The partial evaluator and the backends can easily work around this.