C++ backend: make pairs distinct structs as other tuples
We used to merely typedef pairs as their std::tuple definition to make
it possible a user pair to be defined from a std::tuple returned by a
read function for the runtime lib. This is inconvenient, though, because
it is then no longer possible to declare the user type without giving its
full definition.
So instead this patch adds a specific constructor for that case.