CSV: add the ability to trim each values
Can be used to get rid of spurious blanks after comas, esp. in
ClickHouse format (although that's not in the specifications, Clickhouse
do accept them and some tools rely on this behavior)

Beware that this degrades the generated code for CSV even when not using
the option, due to a limitation of the partial evaluator to be improved
soon.

(cherry picked from commit 7c72f808e5b1a937c237d250a755f2e4fed27739)